CEC2020_p12: evaluate in float64 (float32 overflow made 43% of evaluations -inf)
Settled against the OFFICIAL CEC2020 MATLAB (cec20_func.m, prob_k == 12), not by guessing:

  f = (y1-1).^2 + (y2-1).^2 + (y3-1).^2 - log(y4+1) + (x1-1).^22 + (x2-2).^2 + (x3-3).^2;

* The exponent 22 is NOT our typo -- it is the official formula (very likely a typo in the official
  suite for Kocis & Grossmann's (x1-1)^2, but we reproduce the published benchmark verbatim;
  deviating would make our numbers incomparable to every published CEC2020 result).
* The real defect was ours: base.py casts X to float32, and (100-1)**22 = 8.0e43 is finite in
  float64 (as in MATLAB, which computes in double) but OVERFLOWS float32 (max 3.4e38) to -inf.
  Evaluating this objective in float64 takes finite evaluations from 57% -> 99.2%.
* The integer bounds are NOT changed from -0.51 to -0.49. The official xmin12 is
  [0,0,0,-0.51,-0.51,-0.51,-0.51]; round(-0.51) = -1 -> log(y4+1) = log(0) = -inf occurs in the
  official MATLAB too. The residual 0.8% non-finite is a property of the published problem, not a
  bug here, and "fixing" it would silently make our p12 a different problem.

Truss200D excluded on Rosen's explicit order. Recorded honestly: the reported cause (unbounded FEM
solve, >30 min single eval) did NOT reproduce -- 60 LHS rows evaluated post-fix, none over 5 s. The
hang most likely predates the double-scaling fix (scaling=False). Truss120D is unaffected, stays in.

+++ bocode/opt_problems/cec2020_rw/CEC2020_p1_20.py
@@ -962,7 +962,21 @@ class CEC2020_p12(BenchmarkProblem):
 
         if scaling:
             X = super().scale(X)
-        X = X.numpy()
+        # FLOAT64, deliberately. base.py casts every X to float32, but this objective contains
+        # (x1-1)**22 and x1 ranges to 100: (100-1)**22 = 8.0e43, which is FINITE in float64 (as in the
+        # official MATLAB, which computes in double) but OVERFLOWS float32 (max 3.4e38) to -inf. That
+        # overflow -- not the formula -- made 43% of evaluations non-finite.
+        #
+        # The exponent 22 is NOT our typo: the official CEC2020 cec20_func.m, prob_k == 12, reads
+        #     f = (y1-1).^2 + (y2-1).^2 + (y3-1).^2 - log(y4+1) + (x1-1).^22 + (x2-2).^2 + (x3-3).^2;
+        # It is very likely a typo in the OFFICIAL suite for Kocis & Grossmann's (x1-1)^2, but we
+        # reproduce the official benchmark verbatim -- deviating would make our numbers incomparable
+        # to every published CEC2020 result.
+        #
+        # Likewise the integer bounds stay at -0.51 (official: xmin12 = [0,0,0,-0.51,-0.51,-0.51,-0.51]).
+        # They are NOT changed to -0.49: round(-0.51) = -1 -> log(y4+1) = log(0) = -inf occurs in the
+        # official MATLAB too, so that sliver is a property of the published problem, not a defect here.
+        X = X.numpy().astype(np.float64)
 
         n_samples = X.shape[0]
